termina todo o fluxo de bill

+++ src/main/java/com/pedroresende/rateiofacil/controllers/UserController.java
@@ -1,9 +1,11 @@
 package com.pedroresende.rateiofacil.controllers;
 
 import com.pedroresende.rateiofacil.controllers.dtos.BillDto;
+import com.pedroresende.rateiofacil.controllers.dtos.BillDtoWithitems;
 import com.pedroresende.rateiofacil.controllers.dtos.CreationUserDto;
 import com.pedroresende.rateiofacil.controllers.dtos.ResponseDto;
 import com.pedroresende.rateiofacil.controllers.dtos.UserDto;
+import com.pedroresende.rateiofacil.exceptions.NotAuthorizeUserException;
 import com.pedroresende.rateiofacil.models.entities.Bill;
 import com.pedroresende.rateiofacil.models.entities.User;
 import com.pedroresende.rateiofacil.services.UserService;
@@ -12,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -76,9 +79,11 @@ public class UserController {
    * Mapeamento da rota GET users/{id}.
    */
   @GetMapping("/{id}")
-  public ResponseEntity<UserDto> getById(@PathVariable Long id) {
+  public ResponseEntity<UserDto> getById(@PathVariable Long id,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
     User user = userService.getById(id);
-
+    UserDto userDto = toUserDto(user);
     return ResponseEntity.ok(toUserDto(user));
   }
 
@@ -87,7 +92,9 @@ public class UserController {
    */
   @PutMapping("/{id}")
   public ResponseEntity<ResponseDto<UserDto>> update(@PathVariable Long id,
-      @RequestBody CreationUserDto creationUserDto) {
+      @RequestBody CreationUserDto creationUserDto,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
     User user = userService.update(id, creationUserDto.toEntity());
     ResponseDto<UserDto> responseDto = new ResponseDto<>("Usuário atualizado com sucesso",
         toUserDto(user));
@@ -99,7 +106,9 @@ public class UserController {
    * Mapeamento da rota DELETE users/{id}.
    */
   @DeleteMapping("/{id}")
-  public ResponseEntity<ResponseDto<UserDto>> delete(@PathVariable Long id) {
+  public ResponseEntity<ResponseDto<UserDto>> delete(@PathVariable Long id,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
     User user = userService.delete(id);
     UserDto userDto = toUserDto(user);
     ResponseDto<UserDto> responseDto = new ResponseDto<>("Usuário deletado!", userDto);
@@ -115,8 +124,10 @@ public class UserController {
    * @return mensagem de sucesso e uma instancia de BillDto
    */
   @PostMapping("/{id}/bills")
-  public ResponseEntity<ResponseDto<BillDto>> createCrop(@PathVariable Long id,
-      @RequestBody BillDto billDto) {
+  public ResponseEntity<ResponseDto<BillDto>> createBill(@PathVariable Long id,
+      @RequestBody BillDto billDto,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
     Bill bill = userService.associateBill(id, billDto.toEntity());
 
     BillDto billDtoFromDb = toBillDto(bill);
@@ -130,15 +141,27 @@ public class UserController {
   private BillDto toBillDto(Bill bill) {
 
     if (bill.getUser() == null) {
-      return new BillDto(bill.getId(), null, bill.getEstablishment(), bill.getTotal());
+      return new BillDto(bill.getId(), null, bill.getDate(), bill.getEstablishment(),
+          bill.getTotal());
     }
 
-    return new BillDto(bill.getId(), bill.getUser().getId(), bill.getEstablishment(),
-        bill.getTotal());
+    return new BillDto(bill.getId(), bill.getUser().getId(), bill.getDate(),
+        bill.getEstablishment(), bill.getTotal());
+  }
+
+  private BillDtoWithitems toBillDtoWithitems(Bill bill) {
+    if (bill.getItems() == null) {
+      return new BillDtoWithitems(bill.getId(), bill.getUser().getId(), bill.getDate(),
+          null, bill.getEstablishment(), bill.getTotal());
+    }
+    return new BillDtoWithitems(bill.getId(), bill.getUser().getId(), bill.getDate(),
+        bill.getItems(), bill.getEstablishment(), bill.getTotal());
   }
 
   @GetMapping("/{id}/bills")
-  ResponseEntity<List<BillDto>> getAllBills(@PathVariable Long id) {
+  ResponseEntity<List<BillDto>> getAllBills(@PathVariable Long id,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
     List<Bill> bills = userService.getAllBill(id);
     List<BillDto> billDtos = bills.stream().map(bill -> toBillDto(bill)).toList();
 
@@ -146,9 +169,19 @@ public class UserController {
   }
 
   @GetMapping("/{id}/bills/{billId}")
-  ResponseEntity<BillDto> getBill(@PathVariable Long id, @PathVariable Long billId) {
+  ResponseEntity<BillDto> getBill(@PathVariable Long id, @PathVariable Long billId,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
     Bill bill = userService.getBill(id, billId);
 
     return ResponseEntity.ok(toBillDto(bill));
+  }
+
+  private void validateUserPermission(Long pathId, User authUser) {
+    if (!authUser.getRole().equals("admin")) {
+      if (!authUser.getId().equals(pathId)) {
+        throw new NotAuthorizeUserException();
+      }
+    }
   }
 }
